Move image request queue functions out of ajax.ts and into new image_request_queue.ts.
