+++ src/style/load_sprite.ts
@@ -1,4 +1,5 @@
-import {getJSON, getImage, ResourceType} from '../util/ajax';
+import {getJSON, ResourceType} from '../util/ajax';
+import {getImage} from '../util/image_request_queue';
 
 import browser from '../util/browser';
 import {RGBAImage} from '../util/image';
