+++ src/ui/map.ts
@@ -2,7 +2,8 @@ import {extend, bindAll, warnOnce, uniqueId, isImageBitmap} from '../util/util';
 import browser from '../util/browser';
 import DOM from '../util/dom';
 import packageJSON from '../../package.json' assert {type: 'json'};
-import {getImage, GetImageCallback, getJSON, processImageRequestQueue, installImageQueueThrottleControlCallback, removeImageQueueThrottleControlCallback, ResourceType} from '../util/ajax';
+import {getJSON, ResourceType} from '../util/ajax';
+import {getImage, GetImageCallback, processImageRequestQueue, installImageQueueThrottleControlCallback, removeImageQueueThrottleControlCallback} from '../util/image_request_queue';
 import {RequestManager} from '../util/request_manager';
 import Style, {StyleSwapOptions} from '../style/style';
 import EvaluationParameters from '../style/evaluation_parameters';
